30_server/requirements.txt
django~=5.1.3
djangorestframework~=3.15.2
djangorestframework_simplejwt~=5.3.1
djoser~=2.3.1
